+++ packages/react/src/components/TrackLoop.tsx
@@ -1,7 +1,8 @@
+import { TrackFilter } from '@livekit/components-core';
 import { Track } from 'livekit-client';
 import * as React from 'react';
 import { ParticipantContext } from '../context';
-import { TracksFilter, useTracks } from '../hooks';
+import { useTracks } from '../hooks';
 import { ParticipantTile } from '../prefabs';
 import { cloneSingleChild } from '../utils';
 
@@ -14,7 +15,7 @@ type TrackLoopProps = {
    * Set to `true` if pinned tracks should be included in the participant loop?
    */
   excludePinnedTracks?: boolean;
-  filter?: TracksFilter;
+  filter?: TrackFilter;
   filterDependencies?: Array<any>;
 };