+++ tsup.config.ts
@@ -5,6 +5,7 @@ const defaultOptions: Options = {
   format: ['cjs', 'esm'],
   splitting: false,
   sourcemap: true,
+  treeshake: true,
   // for the type maps to work, we use tsc's declaration-only command
   dts: false,
   clean: true,